services/auth-service/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic-settings>=2.6.0
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
bcrypt>=4.2.0
PyJWT>=2.9.0
email-validator>=2.0.0
